Create Dockerfile & docker-publish CI on tag push Created a dockerfile for the project that is based on the eclipse temurin JDK 21 Noble Ubuntu docker image. The dockerfile compiles and runs the java code for the project.
Additionally I've added the workflow I've created for publishing
docker images to ghcr.io for my git projects.
The workflow triggers when a tag is pushed to the remote repo that
uses semver syntax. An image is built and pushed with the semver tag
and the "latest" tag.

Also added information to the readme on pulling and running the docker image,
as well as some formatting and grammar fixes. A .dockerignore file has
also been included to only copy necessary files into the docker image,
and the .gitignore has been optimised by using a wildcard to catch all
.class files instead of listing them one by one.